Improve naming on different colour setting types
