+++ packages/koenig-lexical/src/components/ui/cards/HeaderCard.jsx
@@ -1,10 +1,10 @@
-import KoenigNestedEditor from '../../KoenigNestedEditor';
 import PropTypes from 'prop-types';
 import React from 'react';
-import {BackgroundImagePicker} from '../BackgroundImagePicker';
-import {Button} from '../Button';
-import {ButtonGroupSetting, ColorPickerSetting, InputSetting, InputUrlSetting, SettingsDivider, SettingsPanel, ToggleSetting} from '../SettingsPanel';
-import {isEditorEmpty} from '../../../utils/isEditorEmpty';
+import { isEditorEmpty } from '../../../utils/isEditorEmpty';
+import KoenigNestedEditor from '../../KoenigNestedEditor';
+import { BackgroundImagePicker } from '../BackgroundImagePicker';
+import { Button } from '../Button';
+import { ButtonGroupSetting, ColorOptionSetting, InputSetting, InputUrlSetting, SettingsDivider, SettingsPanel, ToggleSetting } from '../SettingsPanel';
 
 export const HEADER_COLORS = {
     dark: 'bg-black',
@@ -151,7 +151,7 @@ export function HeaderCard({isEditing,
                         selectedName={size}
                         onClick={handleSizeSelector}
                     />
-                    <ColorPickerSetting
+                    <ColorOptionSetting
                         buttons={colorPickerChildren}
                         label='Style'
                         selectedName={type}
